requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.0.6
django-allauth==0.63.3
django-crispy-forms==2.2
django-heroku==0.3.1
gunicorn==22.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-decouple==3.8
sqlparse==0.5.0
whitenoise==6.7.0
